creating express server for back-end
