+++ src/index.js
@@ -1,24 +1,20 @@
-import express from "express";
-import cors from "cors";
+const express = require('express');
+const cors = require('cors');
+const bodyParser = require('body-parser');
+const BusinessUsercontroller = require('./controllers/BusinessUsercontroller');
+const Usercontroller = require('./controllers/Usercontroller');
+const Reviewscontroller = require('./controllers/Reviewscontroller');
+const { mongoose } = require('./db.js');
 
-// create express app
-const app = express();
 
-// set up CORS
+var app = express();
+app.use(bodyParser.json());
+app.use('/uploads', express.static('uploads'));
 app.use(cors());
 
-// include middleware to enable json body parsing and nested objects
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
 
-// base routes
-app.get("/", (req, res) => {
-  res.send("Welcome to Whatz Hot!");
-});
 
-// routes not found go here
-app.all("*", (req, res) => {
-  res.status(404).send("Oops! Resource not found");
-});
-
-export default app;
+app.listen(3000, () => console.log('server started successfully at port : 3000'));
+app.use('/businessUsers', BusinessUsercontroller);
+app.use('/users', Usercontroller);
+app.use('/reviews', Reviewscontroller);
